Adding the new command and code refactoring
