+++ tests/HighTests/test/campaigns/high/product/1_create_standard_product.js
@@ -1,103 +1,117 @@
-const {selector} = require('../../../globals.webdriverio.js');
+const {AddProductPage} = require('../../../selectors/BO/add_product_page');
+const {AccessPageBO} = require('../../../selectors/BO/access_page');
+var data = require('./../../../datas/product-data');
 
 scenario('Create Standard Product', client => {
   test('should open browser', () => client.open());
-  test('should log in successfully in BO', () => client.signInBO(selector));
-  test('should go to "Catalog"', () => client.waitForExistAndClick(selector.BO.AddProductPage.products_subtab));
-  test('should click on "NEW PRODUCT"', () => client.waitForExistAndClick(selector.BO.AddProductPage.new_product_button));
+  test('should log in successfully in BO', () => client.signInBO(AccessPageBO));
+  test('should go to "Catalog"', () => client.waitForExistAndClick(AddProductPage.products_subtab));
+  test('should click on "NEW PRODUCT"', () => client.waitForExistAndClick(AddProductPage.new_product_button));
 
   scenario('Edit Basic settings', client => {
-    test('should set the "product name"', () => client.setProductName('standard'));
-    test('should set the "Quantity" of product', () => client.setQuantity());
-    // test('should upload the picture one of product', () => client.uploadPicture('1.png'));   // not done
-    // test('should upload the picture two of product', () => client.uploadPicture('2.jpg'));   // not done
-    // test('should upload the picture three of product', () => client.uploadPicture('3.jpg')); // not done
-    test('should click on "CREATE A CATEGORY"', () => client.addCategory());
-    // test('should upload the picture two of product', () => client.uploadPicture('2.jpg'));
-    test('should set the "New category name"', () => client.setCategoryName('standard'));
+    test('should set the "product name"', () => client.waitAndSetValue(AddProductPage.product_name_input, data.standard.name + date_time));
+    test('should set the "Quantity" of product', () => client.waitAndSetValue(AddProductPage.quantity_shortcut_input, "10"));
+    test('should upload the first product picture', () => client.uploadPicture('1.png', AddProductPage.picture));
+    test('should upload the second product picture', () => client.uploadPicture('2.jpg', AddProductPage.picture));
+    test('should upload the third product picture', () => client.uploadPicture('3.jpg', AddProductPage.picture));
+    test('should click on "CREATE A CATEGORY"', () => client.scrollWaitForExistAndClick(AddProductPage.product_create_category_btn, 50));
+    test('should upload the second product picture', () => client.uploadPicture('2.jpg', AddProductPage.picture));
+    test('should set the "New category name"', () => client.waitAndSetValue(AddProductPage.product_category_name_input, data.standard.new_category_name + date_time));
     test('should click on "Create"', () => client.createCategory());
-    test('should remove "HOME" category', () => client.removeHomeCategory());
-    test('should click on "ADD A BRAND"', () => client.addBrand('standard'));
+    test('should remove the "HOME" category', () => client.removeHomeCategory());
+    test('open all category', () => client.openAllCategory());
+    test('should check the existence of the first category Radio button', () => client.checkCategoryRadioButton(1, 1));
+    test('should check the existence of the second category Radio button', () => client.checkCategoryRadioButton(1, 2));
+    test('should check the existence of the third Radio button', () => client.checkCategoryRadioButton(1, 3));
+    test('should check the existence of the fourth Radio button', () => client.checkCategoryRadioButton(2, 1));
+    test('should check the existence of the fifth category Radio button', () => client.checkCategoryRadioButton(2, 2));
+    test('should check the existence of the sixth category Radio button', () => client.checkCategoryRadioButton(2, 3));
+    test('should click on "ADD A BRAND"', () => client.scrollWaitForExistAndClick(AddProductPage.product_add_brand_btn, 50));
     test('should select brand', () => client.selectBrand());
-    test('should click on "ADD RELATED PRODUCT"', () => client.addRelatedProduct('standard'));
+    test('should click on "ADD RELATED PRODUCT"', () => client.waitForExistAndClick(AddProductPage.add_related_product_btn));
     test('should search and add a related product', () => client.searchAndAddRelatedProduct());
     test('should click on "ADD A FEATURE" and select one', () => client.addFeatureHeight('standard'));
-    test('should set "Tax exclude" price', () => client.addProductPriceTaxExcluded());
-    test('should enter the product reference', () => client.addProductReference());
-    test('should set the product "online"', () => client.productOnline());
-  }, 'product/editbasicsettings');
+    test('should set the "Tax exclude" price', () => client.addProductPriceTaxExcluded());
+    test('should set the "Reference"', () => client.waitAndSetValue(AddProductPage.product_reference, data.common.product_reference));
+    test('should set the product "online"', () => client.waitForExistAndClick(AddProductPage.product_online_toggle));
+  }, 'product/product');
 
   scenario('Edit product quantity', client => {
-    test('should click on "Quantities"', () => client.goToProductQuantity());
-    test('should set the "Quantity"', () => client.productQuantity());
-    test('should set the "Minimum quantity for sale"', () => client.minQuantitySale());
-    test('should click on "Deny orders"', () => client.selectAvailabilityPreferences('standard'));
-    test('should set the "label when in stock"', () => client.availableStock());
-    test('should set the "Label when out of stock (and back order allowed)"', () => client.availableOutOfStock());
-    test('should set the "Availability date"', () => client.availabilityDate());
-  }, 'product/editquantity');
-
+    test('should click on "Quantities"', () => client.scrollWaitForExistAndClick(AddProductPage.product_quantities_tab, 50));
+    test('should set the "Quantity"', () => client.waitAndSetValue(AddProductPage.product_quantity_input, data.common.quantity));
+    test('should set the "Minimum quantity for sale"', () => client.waitAndSetValue(AddProductPage.minimum_quantity_sale, data.common.qty_min));
+    test('should click on "Deny orders"', () => client.waitForExistAndClick(AddProductPage.pack_availability_preferences));
+    test('should set the "label when in stock"', () => client.waitAndSetValue(AddProductPage.pack_label_in_stock, data.common.qty_msg_stock));
+    test('should set the "Label when out of stock (and back order allowed)"', () => client.availability());
+    test('should set the "Availability date"', () => client.waitAndSetValue(AddProductPage.pack_availability_date, data.common.qty_date));
+  }, 'product/product');
 
   scenario('Edit product shipping', client => {
-    test('should click on "Shipping"', () => client.goToProductShipping());
-    test('should set the "Width"', () => client.shippingWidth());
-    test('should set the "Height"', () => client.shippingHeight());
-    test('should set the "Depth"', () => client.shippingDepth());
-    test('should set the "Weight"', () => client.shippingWeight());
-    test('should set the "Does this product incur additional shipping costs?"', () => client.shippingCosts());
-    test('should click on "My carrier (Delivery next day!)"', () => client.selectAvailableCarrier());
-  }, 'product/editshipping');
+    test('should click on "Shipping"', () => client.scrollWaitForExistAndClick(AddProductPage.product_shipping_tab, 50));
+    test('should set the "Width"', () => client.waitAndSetValue(AddProductPage.shipping_width, data.common.cwidth));
+    test('should set the "Height"', () => client.waitAndSetValue(AddProductPage.shipping_height, data.common.cheight));
+    test('should set the "Depth"', () => client.waitAndSetValue(AddProductPage.shipping_depth, data.common.cdepth));
+    test('should set the "Weight"', () => client.waitAndSetValue(AddProductPage.shipping_weight, data.common.cweight));
+    test('should set the "Does this product incur additional shipping costs?"', () => client.waitAndSetValue(AddProductPage.shipping_fees, data.common.cadd_ship_coast));
+    test('should click on "My carrier (Delivery next day!)"', () => client.scrollWaitForExistAndClick(AddProductPage.shipping_available_carriers, 50));
+  }, 'product/product');
 
-
-  scenario('Edit product pricing', client => {
-    test('should click on "Pricing"', () => client.goToPricingTab());
-    test('should set the "Price per unit (tax excl.)"', () => client.pricingUnity());
-    test('should set the "Price (tax excl.)"', () => client.pricingWholesale());
-    test('should select the "Priority management"', () => client.pricingPriorities());
-  }, 'product/editpricing');
+  scenario('Edit the product pricing', client => {
+    test('should click on "Pricing"', () => client.scrollWaitForExistAndClick(AddProductPage.product_pricing_tab, 50));
+    test('should set the "Price per unit (tax excl.)"', () => client.waitAndSetValue(AddProductPage.unit_price, data.common.unitPrice));
+    test('should set the "Unit"', () => client.waitAndSetValue(AddProductPage.unity, data.common.unity));
+    test('should set the "Price (tax excl.)"', () => client.waitAndSetValue(AddProductPage.pricing_wholesale, data.common.wholesale));
+    test('should select the "Priority management"', () => client.selectPricingPriorities());
+  }, 'product/product');
 
   scenario('Edit SEO information', client => {
-    test('should click on "SEO"', () => client.goToSEOTab());
-    test('should set the "Meta title"', () => client.metaTitle());
-    test('should set the "Meta description"', () => client.metaDescription());
-    test('should set the "Friendly URL"', () => client.friendlyUrl());
-  }, 'product/editseo');
+    test('should click on "SEO"', () => client.scrollWaitForExistAndClick(AddProductPage.product_SEO_tab, 50));
+    test('should set the "Meta title"', () => client.waitAndSetValue(AddProductPage.SEO_meta_title, data.common.metatitle));
+    test('should set the "Meta description"', () => client.waitAndSetValue(AddProductPage.SEO_meta_description, data.common.metadesc));
+    test('should set the "Friendly URL"', () => client.waitAndSetValue(AddProductPage.SEO_friendly_url, data.common.shortlink));
+  }, 'product/product');
 
   scenario('Edit product options', client => {
-    test('should click on "Options"', () => client.goToOptionsForm());
-    test('should select the "Visibility"', () => client.selectVisibility());
-    test('should click on "Web only (not sold in your retail store)"', () => client.webOnlyVisibility());
+    test('should click on "Options"', () => client.waitForExistAndClick(AddProductPage.product_options_tab));
+    test('should select the "Visibility"', () => client.waitAndSelectByValue(AddProductPage.options_visibility, 'search'));
+    test('should click on "Web only (not sold in your retail store)"', () => client.waitForExistAndClick(AddProductPage.options_online_only));
     test('should select the "Condition"', () => client.selectCondition());
-    test('should set the "ISBN"', () => client.ISBNEntry());
-    test('should set the "EAN-13"', () => client.EAN13Entry());
+    test('should set the "ISBN"', () => client.waitAndSetValue(AddProductPage.options_isbn, data.common.isbn));
+    test('should set the "EAN-13"', () => client.waitAndSetValue(AddProductPage.options_ean13, data.common.ean13));
     test('should set the "UPC"', () => client.UPCEntry());
-    test('should click on "ADD A CUSTOMIZAITION"', () => client.AddCustomFieldButton());
-    test('should set "Label" ,"Type" and "Required"', () => client.createCustomField());
-    test('should click on "ADD A CUSTOMIZAITION"', () => client.AddCustomFieldButton());
-    test('should create new custom field', () => client.newCustomField());
-    test('should click on attach a new file button', () => client.attachNewFile()); // not done
-    test('should add a file', () => client.addFile('image_test.jpg')); // not done
-    test('should select the previous added file', () => client.selectPreviousAddFile());
-  }, 'product/editoptions');
+    test('should click on "ADD A CUSTOMIZAITION FIELD"', () => client.scrollWaitForExistAndClick(AddProductPage.options_add_customization_field_button, 50));
+    test('should set the customization field "Label"', () => client.waitAndSetValue(AddProductPage.options_first_custom_field_label, data.common.personalization.perso_text.name));
+    test('should select the customization field "Type" Text', () => client.waitAndSelectByValue(AddProductPage.options_first_custom_field_type, '1'));
+    test('should click on "Required"', () => client.waitForExistAndClick(AddProductPage.options_first_custom_field_require));
+    test('should click on "ADD A CUSTOMIZAITION"', () => client.scrollWaitForExistAndClick(AddProductPage.options_add_customization_field_button, 50));
+    test('should set the second customization field "Label"', () => client.waitAndSetValue(AddProductPage.options_second_custom_field_label, data.common.personalization.perso_file.name));
+    test('should select the customization field "Type" File', () => client.waitAndSelectByValue(AddProductPage.options_second_custom_field_type, '0'));
+    test('should click on "ATTACH A NEW FILE"', () => client.scrollWaitForExistAndClick(AddProductPage.options_add_new_file_button, 50));
+    test('should add a file', () => client.addFile(AddProductPage.options_select_file, 'image_test.jpg'), 50);
+    test('should set the file "Title"', () => client.waitAndSetValue(AddProductPage.options_file_name, data.common.document_attach.name));
+    test('should set the file "Description" ', () => client.waitAndSetValue(AddProductPage.options_file_description, data.common.document_attach.desc));
+    test('should add the previous added file', () => client.scrollWaitForExistAndClick(AddProductPage.options_file_add_button, 50));
+  }, 'product/product');
 
   scenario('Save Product', client => {
-    test('should click on "SAVE"', () => client.saveProduct());
-    test('should close green validation', () => client.closeGreenValidation());
+    test('should click on "SAVE"', () => client.waitForExistAndClick(AddProductPage.save_product_button));
+    test('should check that the success alert message is well displayed', () => client.waitForExistAndClick(AddProductPage.close_validation_button));
     test('should sign out BO', () => client.signOutBO());
   }, 'product/product');
 
 }, 'product/product', true);
 
-scenario('Check the product in the catalog', client => {
+scenario('Check the product creation', client => {
   test('should open browser', () => client.open());
-  test('should log in successfully in BO', () => client.signInBO(selector));
-  test('should go to "Catalog"', () => client.goToCatalog('standard'));
-  test('should search for product by name', () => client.searchProductByName());
-  test('should check the existance of product name', () => client.checkProductName());
-  test('should check the existance of product reference', () => client.checkProductReference());
-  test('should check the existance of product category', () => client.checkProductCategory());
-  test('should check the existance of product price TE', () => client.checkProductPriceTE());
-  test('should check the existance of product quantity', () => client.checkProductQuantity());
-  test('should check the existance of product status', () => client.checkProductStatus());
-  test('should reset filter', () => client.resetFilter());
-}, 'product/checkproduct', true);
+  test('should log in successfully in BO', () => client.signInBO(AccessPageBO));
+  test('should go to "Catalog"', () => client.goToCatalog());
+  test('should search for product by name', () => client.searchProductByName(data.standard.name + date_time));
+  test('should check the existence of product name', () => client.checkTextValue(AddProductPage.catalog_product_name, data.standard.name + date_time));
+  test('should check the existence of product reference', () => client.checkTextValue(AddProductPage.catalog_product_reference, data.common.product_reference));
+  test('should check the existence of product category', () => client.checkTextValue(AddProductPage.catalog_product_category, data.standard.new_category_name + date_time));
+  test('should check the existence of product price TE', () => client.checkProductPriceTE());
+  test('should check the existence of product quantity', () => client.checkTextValue(AddProductPage.catalog_product_quantity, data.common.quantity));
+  test('should check the existence of product status', () => client.checkTextValue(AddProductPage.catalog_product_online, 'check'));
+  test('should reset filter', () => client.waitForExistAndClick(AddProductPage.catalog_reset_filter));
+}, 'product/check_product', true);
+
