+++ tests/HighTests/test/campaigns/high/product/5_check_default_product_sort.js
@@ -1,6 +1,9 @@
+const {AddProductPage} = require('../../../selectors/BO/add_product_page');
+const {AccessPageBO} = require('../../../selectors/BO/access_page');
+
 scenario('Check default product sort', client => {
   test('should open browser', () => client.open());
-  test('should log in successfully in BO', () => client.signinBO());
-  test('should go to product menu', () => client.goToProductMenu());
-  test('should check default product sort', () => client.getElementID())
+  test('should log in successfully in BO', () => client.signInBO(AccessPageBO));
+  test('should go to "Catalog"', () => client.waitForExistAndClick(AddProductPage.products_subtab));
+  test('should check the default product sort', () => client.getElementID())
 }, 'product/product', true);
