+++ tests/HighTests/test/selectors/BO/access_page.js
@@ -1,0 +1,8 @@
+module.exports = {
+  AccessPageBO:{
+    login_input: '#email',
+    password_inputBO: '#passwd',
+    login_buttonBO: '[name="submitLogin"]',
+    menuBO: '#nav-sidebar'
+  }
+};
