+++ tests/HighTests/test/selectors/BO/catalogpage/Manufacturers/brands.js
@@ -1,0 +1,16 @@
+module.exports = {
+  Brands:{
+    new_brand_button: '//*[@id="page-header-desc-address-new_manufacturer"]',
+    name_input: '//*[@id="name"]',
+    short_desc_textarea: '//*[@id="manufacturer_form"]/div/div[2]/div[2]/div/div/div/div[1]/div/div/div/div/div/div[1]/button',
+    short_desc_source_code_modal: '/html/body/div[5]/div[2]/div[2]/div/textarea',
+    short_desc_source_code_modal_confirmation: '/html/body/div[5]/div[3]/div/div[2]/button',
+    desc_textarea: '//*[@id="manufacturer_form"]/div/div[2]/div[3]/div/div/div/div[1]/div/div/div/div/div/div[1]/button',
+    image_input: '//*[@id="logo"]',
+    meta_title_input: '//*[@id="meta_title_1"]',
+    meta_description_input: '//*[@id="meta_description_1"]',
+    meta_keywords_input:'//*[@id="fieldset_0"]/div[2]/div[7]/div/div/div[1]/div[1]/div/input',   // '//*[@id="manufacturer_form"]/div/div[2]/div[7]/div/div/input',
+    active_button: '//*[@id="manufacturer_form"]/div/div[2]/div[8]/div/span/label[1]',
+    save_button: '//*[@id="manufacturer_form_submit_btn"]'
+  }
+};
