+++ tests/HighTests/test/selectors/BO/catalogpage/Manufacturers/index.js
@@ -1,0 +1,6 @@
+module.exports = Object.assign( {
+  submenu: '//*[@id="subtab-AdminParentManufacturers"]/a'
+},
+  require('./brands'),
+  require('./brands_address')
+);
