+++ tests/HighTests/test/selectors/BO/catalogpage/attribute_submenu.js
@@ -1,0 +1,24 @@
+module.exports = {
+  AttributeSubMenu:{
+    submenu: '//*[@id="subtab-AdminParentAttributesGroups"]/a',
+    add_new_attribute: '//*[@id="page-header-desc-attribute_group-new_attribute_group"]',
+    name_input: '//*[@id="name_1"]',
+    public_name_input: '//*[@id="public_name_1"]',
+    type_select: '//*[@id="group_type"]',
+    save_button: '//*[@id="attribute_group_form_submit_btn"]',
+    search_input: '//*[@id="table-attribute_group"]/thead/tr[2]/th[3]/input',
+    search_button: '//*[@id="submitFilterButtonattribute_group"]',
+    selected_attribute: '//*[@id="table-attribute_group"]/tbody/tr/td[3]',
+    add_value_button: '//*[@id="page-header-desc-attribute-new_value"]',
+    save_and_add: '//*[@id="fieldset_0"]/div[3]/button[2]',
+    save: '//*[@id="attribute_form_submit_btn"]',
+    value_input: '//*[@id="name_1"]',
+    value_action_group_button: '//*[@id="table-attribute"]/tbody/tr[1]/td[5]/div/div/button',
+    delete_value_button: '//*[@id="table-attribute"]/tbody/tr[1]/td[5]/div/div/ul/li/a',
+    group_action_button: '//*[@id="table-attribute_group"]/tbody/tr/td[6]/div/div/button',
+    delete_attribut_button: '//*[@id="table-attribute_group"]/tbody/tr/td[6]/div/div/ul/li[3]/a',
+    update_button: '//*[@id="table-attribute_group"]/tbody/tr/td[6]/div/div/ul/li[1]/a',
+    update_value_button: '//*[@id="table-attribute"]/tbody/tr[1]/td[5]/div/div/a'
+  }
+
+};
