+++ tests/HighTests/test/selectors/BO/catalogpage/feature_submenu.js
@@ -1,0 +1,29 @@
+module.exports = {
+  FeatureSubMenu:{
+    tabmenu: '//*[@id="content"]/div[1]/div/div[2]/a[2]',
+    add_new_feature: '//*[@id="page-header-desc-feature-new_feature"]/div',
+    name_input: '//*[@id="name_1"]',
+    save_button: '//*[@id="feature_form_submit_btn"]',
+    search_input: '//*[@id="table-feature"]/thead/tr[2]/th[3]/input',
+    search_button: '//*[@id="submitFilterButtonfeature"]',
+    selected_feature: '//*[@id="table-feature"]/tbody/tr/td[6]/div/div/a',
+    add_value_button: '//*[@id="page-header-desc-feature_value-new_feature_value"]',
+    value_input: '//*[@id="value_1"]',
+    save_value_button: '//*[@id="feature_value_form_submit_btn"]',
+    select_option: '//*[@id="table-feature"]/tbody/tr/td[6]/div/div/button',
+    update_feature_button: '//*[@id="table-feature"]/tbody/tr/td[6]/div/div/ul/li[1]/a',
+    update_feature_value_button: '//*[@id="table-feature_value"]/tbody/tr/td[3]/div/div/a',
+    delete_feature: '//*[@id="table-feature"]/tbody/tr/td[6]/div/div/ul/li[3]/a',
+    public_name_input: '//*[@id="public_name_1"]',
+    type_select: '//*[@id="group_type"]',
+    selected_attribute: '//*[@id="table-attribute_group"]/tbody/tr/td[3]',
+    save_and_add: '//*[@id="fieldset_0"]/div[3]/button[2]',
+    save: '//*[@id="attribute_form_submit_btn"]',
+    value_action_group_button: '//*[@id="table-attribute"]/tbody/tr[1]/td[5]/div/div/button',
+    delete_value_button: '//*[@id="table-attribute"]/tbody/tr[1]/td[5]/div/div/ul/li/a',
+    group_action_button: '//*[@id="table-attribute_group"]/tbody/tr/td[6]/div/div/button',
+    delete_attribut_button: '//*[@id="table-attribute_group"]/tbody/tr/td[6]/div/div/ul/li[3]/a',
+    update_button: '//*[@id="table-attribute_group"]/tbody/tr/td[6]/div/div/ul/li[1]/a',
+    update_value_button: '//*[@id="table-attribute"]/tbody/tr[1]/td[5]/div/div/a'
+  }
+};
