+++ tests/HighTests/test/selectors/BO/catalogpage/stocksubmenu/index.js
@@ -1,0 +1,5 @@
+module.exports = Object.assign(
+  {},
+  require('./movements'),
+  require('./stock')
+);
