+++ tests/HighTests/test/selectors/BO/catalogpage/stocksubmenu/movements.js
@@ -1,0 +1,6 @@
+module.exports = {
+  Movement:{
+    tabs: '//*[@id="tab"]/li[2]/a',
+    variation: '//*[@id="app"]/div[3]/section/table/tbody/tr[1]/td[4]/span/span'
+  }
+};
