+++ tests/HighTests/test/selectors/BO/catalogpage/stocksubmenu/stock.js
@@ -1,0 +1,24 @@
+module.exports = {
+  Stock:{
+    submenu: '//*[@id="collapse-9"]/li[8]/a',
+    tabs: '//*[@id="tab"]/li[1]/a',
+    first_product_quantity_input: '//*[@id="app"]/div[4]/section/table/tbody/tr[1]/td[8]/form/div/input',
+    first_product_quantity: '//*[@id="app"]/div[4]/section/table/tbody/tr[1]/td[7]',
+    first_product_quantity_modified: '//*[@id="app"]/div[4]/section/table/tbody/tr[1]/td[7]/span',
+    second_product_quantity_input: '//*[@id="app"]/div[4]/section/table/tbody/tr[2]/td[8]/form/div/input',
+    second_product_quantity: '//*[@id="app"]/div[4]/section/table/tbody/tr[2]/td[7]',
+    second_product_quantity_modified: '//*[@id="app"]/div[4]/section/table/tbody/tr[2]/td[7]/span',
+    third_product_quantity_input: '//*[@id="app"]/div[4]/section/table/tbody/tr[3]/td[8]/form/div/input',
+    third_product_quantity: '//*[@id="app"]/div[4]/section/table/tbody/tr[3]/td[7]',
+    third_product_quantity_modified: '//*[@id="app"]/div[4]/section/table/tbody/tr[3]/td[7]/span',
+    save_third_product_quantity_button: '//*[@id="app"]/div[4]/section/table/tbody/tr[3]/td[8]/form/button',
+    fourth_product_quantity_input: '//*[@id="app"]/div[4]/section/table/tbody/tr[4]/td[8]/form/div/input',
+    fourth_product_quantity: '//*[@id="app"]/div[4]/section/table/tbody/tr[4]/td[7]',
+    fourth_product_quantity_modified: '//*[@id="app"]/div[4]/section/table/tbody/tr[4]/td[7]/span',
+    save_fourth_product_quantity_button: '//*[@id="app"]/div[4]/section/table/tbody/tr[4]/td[8]/form/button',
+    group_apply_button: '//*[@id="app"]/div[4]/section/div/div[2]/button',
+    add_quantity_button: '//*[@id="app"]/div[4]/section/table/tbody/tr[3]/td[8]/form/div/div/span[1]',
+    remove_quantity_button: '//*[@id="app"]/div[4]/section/table/tbody/tr[4]/td[8]/form/div/div/span[2]',
+    success_panel: '//*[@id="growls"]'
+  }
+};
