+++ tests/HighTests/test/selectors/BO/customers/addresses.js
@@ -1,0 +1,21 @@
+module.exports = {
+  Addresses:{
+    addresses_menu: '//*[@id="subtab-AdminAddresses"]/a',
+    new_address_button:'//*[@id="page-header-desc-address-new_address"]',
+    email_input: '//*[@id="email"]',
+    id_number_input: '//*[@id="dni"]',
+    address_alias_input: '//*[@id="alias"]',
+    first_name_input: '//*[@id="firstname"]',
+    last_name_input: '//*[@id="lastname"]',
+    company: '//*[@id="company"]',
+    VAT_number_input: '//*[@id="vat_number"]',
+    address_input: '//*[@id="address1"]',
+    address_second_input: '//*[@id="address2"]',
+    zip_code_input: '//*[@id="postcode"]',
+    city_input: '//*[@id="city"]',
+    country_input: '//*[@id="id_country"]',
+    phone_input: '//*[@id="phone"]',
+    other_input: '//*[@id="other"]',
+    save_button:'//*[@id="address_form_submit_btn"]'
+  }
+}
