+++ tests/HighTests/test/selectors/BO/customers/customer.js
@@ -1,0 +1,15 @@
+module.exports = {
+  Customer:{
+    customer_menu: '//*[@id="subtab-AdminParentCustomer"]/a',
+    new_customer_button: '//*[@id="page-header-desc-customer-new_customer"]',
+    social_title_button: '//*[@id="gender_1"]',
+    first_name_input: '//*[@id="firstname"]',
+    last_name_input: '//*[@id="lastname"]',
+    email_address_input: '//*[@id="email"]',
+    password_input: '//*[@id="passwd"]',
+    days_select: '//*[@id="fieldset_0"]/div[2]/div[6]/div/div/div[1]/select',
+    month_select: '//*[@id="fieldset_0"]/div[2]/div[6]/div/div/div[2]/select',
+    years_select: '//*[@id="fieldset_0"]/div[2]/div[6]/div/div/div[3]/select',
+    save_button: '//*[@id="customer_form_submit_btn"]'
+  }
+};
