+++ tests/HighTests/test/selectors/BO/customers/index.js
@@ -1,0 +1,5 @@
+module.exports =Object.assign(
+  {},
+  require('./addresses'),
+  require('./customer')
+  );
