+++ tests/HighTests/test/selectors/BO/index.js
@@ -1,0 +1,11 @@
+module.exports = Object.assign({},
+  require('./catalogpage'),
+  require('./customers'),
+  require('./access_page'),
+  require('./add_product_page'),
+  require('./module_page'),
+  require('./onboarding'),
+  require('./order_page'),
+  require('./product_list')
+);
+
