+++ tests/HighTests/test/selectors/BO/onboarding.js
@@ -1,0 +1,7 @@
+module.exports = {
+  OnBoarding:{
+    popup: '.onboarding-popup',
+    popup_close_button: '/html/body/div[1]/div/div/div[3]/button[1]',
+    stop_button: '.onboarding-button-stop'
+  }
+};
