+++ tests/HighTests/test/selectors/BO/product_list.js
@@ -1,0 +1,7 @@
+module.exports = {
+  ProductList:{
+    first_product_id:'//*[@id="product_catalog_list"]/div[2]/div/table/tbody/tr[1]/td[1]/div/label',
+    second_product_id:'//*[@id="product_catalog_list"]/div[2]/div/table/tbody/tr[2]/td[1]/div/label',
+    third_product_id:'//*[@id="product_catalog_list"]/div[2]/div/table/tbody/tr[3]/td[1]/div/label'
+  }
+};
