+++ tests/HighTests/test/selectors/FO/access_page.js
@@ -1,0 +1,11 @@
+module.exports = {
+ AccessPageFO:{
+  sign_in_button: 'div.user-info > a',
+  login_input: '//*[@id="login-form"]/section/div[1]/div[1]/input',
+  password_inputFO: '//*[@id="login-form"]/section/div[2]/div[1]/div/input',
+  login_button: '//*[@id="login-form"]/footer/button',
+  sign_out_button: '.logout',
+  logo_home_page: '.logo.img-responsive',
+  product_list_button: '//*[@id="content"]/section/a'
+ }
+};
