+++ tests/HighTests/test/selectors/FO/add_account_page.js
@@ -1,0 +1,8 @@
+module.exports = {
+  create_button: '[data-link-action="display-register-form"]',
+  firstname_input: '[name="firstname"]',
+  lastname_input: '[name="lastname"]',
+  email_input: '[name="email"]',
+  password_input: '[name="password"]',
+  save_account_button: '[data-link-action="save-customer"]'
+};
