+++ tests/HighTests/test/selectors/FO/layer_cart_page.js
@@ -1,0 +1,7 @@
+module.exports = {
+  layer_cart: '//div[@id="blockcart-modal" and @style="display: block;"]',
+  name_details: '//div[@id="blockcart-modal"]/div/div/div[2]/div/div[1]/div/div[2]/h6',
+  price_details: '//div[@id="blockcart-modal"]/div/div/div[2]/div/div[1]/div/div[2]/p[1]',
+  quantity_details: '//div[@id="blockcart-modal"]/div/div/div[2]/div/div[1]/div/div[2]/p[2]',
+  command_button: '//*[@id="blockcart-modal"]/div/div/div[2]/div/div[2]/div/div/a'
+};
