+++ tests/HighTests/test/selectors/FO/product_page.js
@@ -1,0 +1,3 @@
+module.exports = {
+  title: '//*[@id="main"]/div[1]/div[2]/h1'
+};
