+++ tests/HighTests/test/selectors/FO/search_product_page.js
@@ -1,0 +1,14 @@
+module.exports = {
+  category_list: '//*[@id="left-column"]/div[1]/ul/li[2]/ul/li',
+  second_category_name: '//*[@id="left-column"]/div[1]/ul/li[2]/ul/li[2]/a',
+  product_search_input: '.ui-autocomplete-input',
+  product_search_button: '.material-icons.search',
+  product_result_name: '.h3.product-title > a',
+  product_result_price: '[itemprop="price"]',
+  attribut_name: '//*[@id="add-to-cart-or-refresh"]/div[1]/div/span',
+  feature_name: '//*[@id="product-details"]/section/dl/dt',
+  feature_value: '//*[@id="product-details"]/section/dl/dd',
+  attribut_value_1: '//*[@id="add-to-cart-or-refresh"]/div[1]/div/ul/li[1]/label/span',
+  attribut_value_2: '//*[@id="add-to-cart-or-refresh"]/div[1]/div/ul/li[2]/label/span',
+  attribut_value_3: '//*[@id="add-to-cart-or-refresh"]/div[1]/div/ul/li[3]/label/span'
+};
