+++ tests/HighTests/test/selectors/index.js
@@ -1,0 +1,6 @@
+module.exports = Object.assign(
+  {},
+  require('./BO'),
+  require('./FO')
+);
+
